requirements.txt
matrix-nio
aiofiles
markdown
requests
beautifulsoup4